mystic-tarot-crew/requirements.txt
# CrewAI and Core Dependencies
crewai>=0.28.0
crewai-tools>=0.2.0

# API Integration
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing
typing-extensions>=4.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.0.0

# Logging & Monitoring
structlog>=23.0.0
colorlog>=6.0.0

# Optional: Enhanced functionality
pydantic>=2.0.0
rich>=13.0.0